requirements.txt
termcolor
markdown
pyyaml
typing_extensions
matrix-nio[e2e]
geoip2==4.6.0
Mastodon.py==1.8.0
maxminddb==2.2.0
